Simplify input point handling
I've gone back and forth in various plugins with whether a single input point should always be picked, and a dedicated object be kept as reference, or a new input point should be picked for each step; lol.

Now trying to move responsibilities back onto InputPoint instead of having a separate @point variable.